backend/chat-agent/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
pydantic-ai>=0.0.7
supabase>=2.0.0
joblib>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0
openai>=1.0.0